feat: implement loading context and refactor loading component
