--- client/src/components/loading/useLoading.js
@@ -1,4 +1,0 @@
-import { useContext } from "react";
-import { LoadingContext } from "./LoadingContext.jsx";
-
-export const useLoading = () => useContext(LoadingContext);
