+++ client/src/hooks/useLoading.js
@@ -1,0 +1,7 @@
+import { useContext } from "react";
+
+import { LoadingContext } from "../contexts/LoadingContext";
+
+
+// Custom hook to use the LoadingContext
+export const useLoading = () => useContext(LoadingContext);
